+++ README.md
@@ -1,2 +1,11 @@
-# container-apps-message-processor
-Create Azure Service Bus message processor and deploy them in Azure Container Apps
+# Message Processor Deployment
+
+This project contains three folders for deploying a message processor. Each deployment receives messages from a service bus and logs them in the console.
+
+## Deployment Options
+
+1. [fabrikam-functions](./fabrikam-functions/README.md): Deploys an Azure Function in an Azure Container Apps Environment.
+
+2. [fabrikam-container-app](./fabrikam-container-app/README.md): Deploys a Java application using Quarkus in an Azure Container App.
+
+3. [fabrikam-container-app-job](./fabrikam-container-app-job/README.md): Deploys a job written using the Azure SDK for Java in an Azure Container App.
