+++ fabrikam-container-app-job/infra/bicep/main.bicep
