+++ fabrikam-container-app/.dockerignore
