+++ fabrikam-container-app/infra/bicep/main.bicep
