+++ fabrikam-container-app/src/main/docker/Dockerfile.jvm
