+++ fabrikam-container-app/src/main/docker/Dockerfile.legacy-jar
