+++ fabrikam-container-app/src/main/docker/Dockerfile.native-micro
